requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
html2image==2.0.4.3
idna==3.6
multidict==6.0.4
outcome==1.3.0.post0
pycparser==2.21
PySocks==1.7.1
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.1.0
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.4
